+++ app/src/main/java/com/example/bank/bankapp/model/Payment.kt
@@ -4,5 +4,5 @@ class Payment (
     val title : String,
     val desc : String,
     val date : String,
-    val value : Int
+    val value : Double
 )
